Fix <rdar://problem/5365030> calling dataWithPDFInsideRect on an SVG with a gradient crashes (14780)
Reviewed by Anders.

When drawing directly to PDF CG may delay the use of the gradient function until outside our
standard drawing path, which in turn could let us invalidate the caches before they were used.

To work around this we now store the cached stops in a RefCounted object, so that we can ensure
that cache exists as long as required.


git-svn-id: svn://svn.chromium.org/blink/trunk@27781 bbb929c8-8fbe-4397-9dbb-9b2b20218538