Improve performance of CTI on windows.
Reviewed by Geoff Garen

Currently on platforms where the compiler doesn't allow us to safely
index relative to the address of a parameter we need to actually
provide a pointer to CTI runtime call arguments.  This patch improves
performance in this case by making the CTI logic for restoring this
parameter much less conservative by only resetting it before we actually
make a call, rather than between each and every SF bytecode we generate
code for.

This results in a 3.6% on the v8 benchmark when compiled with MSVC.


git-svn-id: svn://svn.chromium.org/blink/trunk@37138 bbb929c8-8fbe-4397-9dbb-9b2b20218538