[GTK] Fix crash with LayoutTests/fast/loader/frame-creation-removal.html
Call WebCore::Frame::init after the frame has been given
a name and been added to the FrameTree of the parent.

Removing the call to init is fine as FrameLoaderClientGtk::createFrame
is the only user of webkit_web_frame_init_with_web_view and is already
calling WebCore::Frame::init.

git-svn-id: svn://svn.chromium.org/blink/trunk@39278 bbb929c8-8fbe-4397-9dbb-9b2b20218538