[GIO] Attach the ResourceHandle to the GObject and get it from the GObject
https://bugs.webkit.org/show_bug.cgi?id=23116 discussed with Benjamin Otte

Sometimes it is not possible to cancel a pending GIO operation and the
callback might be invoked with a pointer to an already destructed
ResourceHandle. To avoid this issue we will attach the ResourceHandle
to the GObject* and in the callback attempt to get the ResourceHandle
from the GObject*. If no ResourceHandle is attached we silently fail
as the cleanup has been already done.

git-svn-id: svn://svn.chromium.org/blink/trunk@39857 bbb929c8-8fbe-4397-9dbb-9b2b20218538