<rdar://problem/7006959> 'Save as…’ does not work in Mail
Reviewed by Timothy Hatcher.

* WebView/WebFrameView.mm:
(-[WebFrameView initWithFrame:]): Call _needMailThreadWorkaroundIfCalledOffMainThread rather than
needMailThreadWorkaround, as the latter is intended to be used at the point which a workaround would
be applied and thus contains a main thread check.  Since -initWithFrame: is called on the main thread,
this was causing us to not switch from exception-throwing to logging for the thread violation behavior.

git-svn-id: svn://svn.chromium.org/blink/trunk@45386 bbb929c8-8fbe-4397-9dbb-9b2b20218538