Fix <rdar://problem/6954546> Crashes in WebCore::RenderPart::setWidget below FrameLoader::loadPlugin.
Reviewed by Dan Bernstein.

The calls to FrameLoader::requestObject within RenderPartObject::updateWidget can result in a plug-in
being initialized.  This can run cause arbitrary JavaScript to run and may result in the RenderPartObject
being detached from the render tree and destroyed, causing a crash like <rdar://problem/6954546>.

No test is possible until <rdar://problem/7090444> is fixed

* rendering/RenderPartObject.cpp:
(WebCore::RenderPartObject::updateWidget): Ensure that we stay alive for the duration of plug-in initialization.

git-svn-id: svn://svn.chromium.org/blink/trunk@48359 bbb929c8-8fbe-4397-9dbb-9b2b20218538