Speed up access to history items by caching date computation.
Patch by Darin Adler <darin@apple.com> on 2009-09-25
Reviewed by Steve Falkenburg.

* History/WebHistory.mm:
(getDayBoundaries): Refactored from timeIntervalForBeginningOfDay.
Returns the beginning of the day that the passed time is within and
the beginning of the next day.
(beginningOfDay): Added. Uses getDayBoundaries so it can be fast for
multiple dates within the same day, which is the common case.
(dateKey): Added. Calls beginningOfDay and converts to an integer.
(WebHistory::findKey): Changed to call dateKey instead of
timeIntervalForBeginningOfDay.



git-svn-id: svn://svn.chromium.org/blink/trunk@48770 bbb929c8-8fbe-4397-9dbb-9b2b20218538