Unreviewed documentation fix from David Boddie (Qt Doc Team)
Patch by Kenneth Rohde Christiansen <kenneth@webkit.org> on 2009-10-26
Removes the check around the RenderHints property documentation
that was clearly added to synchronize the source and header files
when the #if !defined(Q_OS_SYMBIAN) guards was added to the
property.

The documentation has also been updated to ensure that Symbian
users know that there is no actual RenderHints property on their
platform.

* Api/qwebview.cpp:

git-svn-id: svn://svn.chromium.org/blink/trunk@50099 bbb929c8-8fbe-4397-9dbb-9b2b20218538