Export fastMalloc, fastCalloc, fastRealloc and fastFree on GCC/Unix
https://bugs.webkit.org/show_bug.cgi?id=30769

When using -fvisibility=hidden to hide all internal symbols by default
the malloc symbols will be hidden as well. For memory instrumentation
it is needed to provide an instrumented version of these symbols and
override the normal routines and by changing the visibility back to
default this becomes possible.

The only other solution would be to use system malloc instead of the
TCmalloc implementation but this will not allow to analyze memory
behavior with the default allocator.

* wtf/FastMalloc.h: Define WTF_FAST_MALLOC_EXPORT for GCC and !darwin

git-svn-id: svn://svn.chromium.org/blink/trunk@50204 bbb929c8-8fbe-4397-9dbb-9b2b20218538