[Qt] Implement load error pages support for Qt's DRT. https://bugs.webkit.org/show_bug.cgi?id=31509
Reviewed by Kenneth Christiansen.

For now, it will not be a default feature, and layout tests
that want to make use of this have to explicitily call
'handleErrorPages();' for the test source.

Any of the other DumpRenderTree's (mac, win and gtk)
support handling error pages. Qt's will be the first.

* DumpRenderTree/qt/DumpRenderTree.cpp:
(WebCore::WebPage::supportsExtension):
(WebCore::WebPage::extension):
* DumpRenderTree/qt/LayoutTestControllerQt.cpp:
(LayoutTestController::reset):
* DumpRenderTree/qt/LayoutTestControllerQt.h:
(LayoutTestController::shouldHandleErrorPages):
(LayoutTestController::handleErrorPages):

git-svn-id: svn://svn.chromium.org/blink/trunk@51037 bbb929c8-8fbe-4397-9dbb-9b2b20218538