WebCore: [Qt] JavaScript prompt is currently broken. https://bugs.webkit.org/show_bug.cgi?id=30914
Patch by Holger Hans Peter Freyther <zecke@selfish.org> on 2010-01-26
Reviewed by Simon Hausmann.

Remove the manual test case in favor of an automated
test case in WebKit/qt/tests/qwebpage.

* manual-tests/qt/java-script-prompt.html: Removed.

WebKit/qt: [Qt] JavaScript prompt is currently broken
https://bugs.webkit.org/show_bug.cgi?id=30914

Patch by Holger Hans Peter Freyther <zecke@selfish.org> on 2010-01-26
Reviewed by Simon Hausmann.

In r52152 a patch was landed to convert a null QString
to an empty WebCore::String in case the prompt was accepted
but the default implementation returned the null QString.

The patch tried to avoid assign to result twice and
was not checking the QString if it is null but the default
value. This lead to always returning an empty string on
successful prompts. Fix it by checking the variable 'x'
for isNull.

The manual test case used didn't cover the case of non
empty input, replace it with an automatic test case that
should cover all cases.

* WebCoreSupport/ChromeClientQt.cpp:
(WebCore::ChromeClientQt::runJavaScriptPrompt): Fix the bug.
* tests/qwebpage/tst_qwebpage.cpp: Add automatic test case
(JSPromptPage::JSPromptPage):
(JSPromptPage::javaScriptPrompt):
(tst_QWebPage::testJSPrompt):

git-svn-id: svn://svn.chromium.org/blink/trunk@53847 bbb929c8-8fbe-4397-9dbb-9b2b20218538