WebCore: Based on a pacth from Nick Jong.
Reviewed by Simon Fraser.

Improve selection in multi-column blocks when hitting points above or
below a column rect.

Test: fast/multicol/hit-test-above-or-below.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::adjustPointToColumnContents): If the point lies
within the horizontal range for a column, constrain it to the column (if
it is above) or the next column (if it is below).

LayoutTests: Added test for hit-testing points above or below the column rect.

Reviewed by Simon Fraser.

* fast/multicol/hit-test-above-or-below-expected.txt: Added.
* fast/multicol/hit-test-above-or-below.html: Added.



git-svn-id: svn://svn.chromium.org/blink/trunk@55546 bbb929c8-8fbe-4397-9dbb-9b2b20218538