protocolHostAndPortEquals host check makes a wrong assumption https://bugs.webkit.org/show_bug.cgi?id=37777
Reviewed by Alexey Proskuryakov.

WebCore:

The host check assumed that both host started at the same position. This is true
if both URL are the same but sometimes one has credential and the other does not.
In this case, the method would compare invalid positions.

Test: http/tests/appcache/credential-url.html

* platform/KURL.cpp:
(WebCore::protocolHostAndPortAreEqual):
* platform/KURLGoogle.cpp:
(WebCore::protocolHostAndPortAreEqual):
Fix the host check to take both URL's credential into account.

LayoutTests:

Add a check that the URL with credential matches the current URL.

* http/tests/appcache/credential-url-expected.txt: Added.
* http/tests/appcache/credential-url.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@58219 bbb929c8-8fbe-4397-9dbb-9b2b20218538