Don't initialize plug-ins until allowed by the page https://bugs.webkit.org/show_bug.cgi?id=42033
Reviewed by Sam Weinig.

WebCore:

* WebCore.exp.in:
Export Document::addMediaCanStartListener and Document::removeMediaCanStartListener.

WebKit2:

* WebProcess/Plugins/Netscape/NetscapePlugin.cpp:
(WebKit::NetscapePlugin::paint):
(WebKit::NetscapePlugin::geometryDidChange):
Assert that the plug-in is started.

* WebProcess/Plugins/PluginView.cpp:
(WebKit::PluginView::PluginView):
Initialize m_isWaitingUntilMediaCanStart.

(WebKit::PluginView::~PluginView):
If necessary, remove the plug-in view as a MediaCanStartListener.

(WebKit::PluginView::initializePlugin):
If we're not allowed to initialize the plug-in, add the plug-in view as a MediaCanStartListener.

(WebKit::PluginView::paint):
Check that the plug-in is initialized.

(WebKit::PluginView::viewGeometryDidChange):
Ditto.

(WebKit::PluginView::mediaCanStart):
Initialize the plug-in.

* WebProcess/Plugins/PluginView.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@63053 bbb929c8-8fbe-4397-9dbb-9b2b20218538