[Qt] Change default font from sans-serif to serif
Reviewed by Simon Hausmann.

This matches what other ports and browser engines do.

We now also use QFont's defaultFontFamily() to resolve the
correct fonts for each platform. For Qt >= 4.7 we can take
advantage of the new enums for generic font families.

The spatial-navigation test had to be changed due to the
fragile nature of the down-action. By changing the font we
hit the case where the two down actions were not enough to
scroll the link into view.

git-svn-id: svn://svn.chromium.org/blink/trunk@64114 bbb929c8-8fbe-4397-9dbb-9b2b20218538