AX: Misspellings not shown in AXAttributedStringForTextMarkerRange when selection is on word https://bugs.webkit.org/show_bug.cgi?id=41817
Reviewed by David Kilzer.

WebCore: 

Change how misspelled words are identified for accessibility. No longer rely on the 
cached markers in Document. Calculate on the fly, which is more accurate in a variety of 
circumstances (like when selection has not moved across words, of if the cursor is in the middle
of a misspelled word). 

Test: platform/mac/accessibility/attributed-string-includes-misspelled-with-selection.html

* accessibility/mac/AccessibilityObjectWrapper.mm:
(AXAttributeStringSetSpelling):
(AXAttributedStringAppendText):
(-[AccessibilityObjectWrapper doAXAttributedStringForTextMarkerRange:]):

LayoutTests: 

* platform/mac/accessibility/attributed-string-includes-misspelled-with-selection-expected.txt: Added.
* platform/mac/accessibility/attributed-string-includes-misspelled-with-selection.html: Added.



git-svn-id: svn://svn.chromium.org/blink/trunk@64149 bbb929c8-8fbe-4397-9dbb-9b2b20218538