Don't use bufferIsLargeEnoughToContain for Vectors with variable sized elements https://bugs.webkit.org/show_bug.cgi?id=46109
Reviewed by Adam Roben.

Add a new VectorArgumentCoder class template, with specializations based on whether the
element is fixed size or not. Then update the ArgumentEncoder<Vector<T>> specialization to choose the
right VectorArgumentCoder specialization. To determine this, we use the "IsArithmetic" type trait,
which holds true for all integer types and all floating point types.

* Platform/CoreIPC/ArgumentCoders.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@67874 bbb929c8-8fbe-4397-9dbb-9b2b20218538