https://bugs.webkit.org/show_bug.cgi?id=46314.
Reviewed by Simon Fraser.

The RenderView's block-flow should match the root element's block-flow.  When we create (or re-create) the style for the document, take
the block-flow from the document element's style and use it if it's available.
        
Also patch styleDidChange so that if the block-flow changes dynamically after we have already made the document style, we'll propagate
the change back up to the RenderView and have it do a relayout.

No tests possible yet, since you can't see the document's style anywhere, and block-flow isn't doing anything to affect layout
yet.

* css/CSSStyleSelector.cpp:
(WebCore::CSSStyleSelector::styleForDocument):
* rendering/RenderBox.cpp:
(WebCore::RenderBox::styleDidChange):



git-svn-id: svn://svn.chromium.org/blink/trunk@68157 bbb929c8-8fbe-4397-9dbb-9b2b20218538