Crash at lineBreakExistsAtPosition + 125 <rdar://problem/7028809> https://bugs.webkit.org/show_bug.cgi?id=46770
Reviewed by Darin Adler.

lineBreakExistsAtPosition is called from InsertTextCommand::input,
where we compute the downstream position of the endingSelection().
Downstream can return the original position, that comes from a VisibleSelection
but there is no guarantee that its renderer is still there. Everywhere we dereference
a renderer we check if it's null.
        
There is no regression test.
        
* editing/htmlediting.cpp:
(WebCore::lineBreakExistsAtPosition): Added check that the renderer is not null
before dereferencing it.



git-svn-id: svn://svn.chromium.org/blink/trunk@68727 bbb929c8-8fbe-4397-9dbb-9b2b20218538